restricted pages added

+++ base/views.py
@@ -1,32 +1,41 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.http import HttpResponse 
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
 from django.http import HttpResponse
 from .models import Room, Topic
 from .forms import RoomForm
 # Create your views here.
 
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login
+from django.contrib import messages
+
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
 
         try:
             user = User.objects.get(username=username)
-        except:
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                login(request, user)
+                return redirect('home')
+            else:
+                messages.error(request, 'Password is not correct')
+        except User.DoesNotExist:
             messages.error(request, 'User does not exist')
 
-        user = authenticate(request, username=username, password=password)
-
-        if user is not None:
-            login(request, user)
-            return redirect('home')
-        else:
-            messages.error(request, 'Password is not correct')
     context = {}
     return render(request, 'base/login_register.html', context)
+
 
 def logoutUser(request):
     logout(request)
@@ -51,6 +60,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -62,10 +72,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('You are not the host')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid:
@@ -75,8 +89,13 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not the host')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
